backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
pymongo==4.6.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
email-validator
pytz==2023.3
requests==2.31.0
gradio_client==2.0.3
Pillow==10.1.0
numpy==1.26.2